Shaderc-rs & Shaderc-sys SCHISM
Introducing shaderc-sys

The new sys crate includes all of the original ffi and build functionality.  Build instructions are left unchanged, but only the setup phase remains valid due to transitive requirements of the underlying sys crate.

Within lib, the only change was to rename the ffi crate to scs for "shaderc-sys".